chore(remix): Move server and lib to app directory
