+++ app/server/lib/db/cli.ts
