+++ app/server/zod/common/RecordSoft.ts
