+++ app/server/zod/plate/common/RichText.ts
