+++ app/server/zod/plate/common/RichTextBase.ts
