+++ app/server/zod/plate/common/Text.ts
